+++ litebox_shim_linux/src/lib.rs
@@ -218,24 +218,24 @@ const SYS_MMAP: i64 = 9;
 const SYS_OPENAT: i64 = 257;
 
 /// Entry point for the syscall handler
-pub fn syscall_entry(request: SyscallRequest<Platform>) -> isize {
+pub fn syscall_entry(request: SyscallRequest<Platform>) -> i64 {
     match request {
         SyscallRequest::Read { fd, buf, count } => {
             let Ok(count) = isize::try_from(count) else {
-                return Errno::EINVAL.as_neg() as isize;
+                return i64::from(Errno::EINVAL.as_neg());
             };
             buf.mutate_subslice_with(..count, |user_buf| {
                 // TODO: use kernel buffer to avoid page faults
                 syscalls::file::sys_read(fd, user_buf, None).map_or_else(
-                    |e| e.as_neg() as isize,
+                    |e| i64::from(e.as_neg()),
                     #[allow(clippy::cast_possible_wrap)]
-                    |size| size as isize,
+                    |size| size as i64,
                 )
             })
-            .unwrap_or(Errno::EFAULT.as_neg() as isize)
+            .unwrap_or(i64::from(Errno::EFAULT.as_neg()))
         }
         SyscallRequest::Close { fd } => {
-            syscalls::file::sys_close(fd).map_or_else(Errno::as_neg, |()| 0) as isize
+            i64::from(syscalls::file::sys_close(fd).map_or_else(Errno::as_neg, |()| 0))
         }
         SyscallRequest::Pread64 {
             fd,
@@ -244,17 +244,17 @@ pub fn syscall_entry(request: SyscallRequest<Platform>) -> isize {
             offset,
         } => {
             let Ok(count) = isize::try_from(count) else {
-                return Errno::EINVAL.as_neg() as isize;
+                return i64::from(Errno::EINVAL.as_neg());
             };
             buf.mutate_subslice_with(..count, |user_buf| {
                 // TODO: use kernel buffer to avoid page faults
                 syscalls::file::sys_pread64(fd, user_buf, offset).map_or_else(
-                    |e| e.as_neg() as isize,
+                    |e| i64::from(e.as_neg()),
                     #[allow(clippy::cast_possible_wrap)]
-                    |size| size as isize,
+                    |size| size as i64,
                 )
             })
-            .unwrap_or(Errno::EFAULT.as_neg() as isize)
+            .unwrap_or(i64::from(Errno::EFAULT.as_neg()))
         }
         SyscallRequest::Mmap {
             addr,
@@ -265,10 +265,10 @@ pub fn syscall_entry(request: SyscallRequest<Platform>) -> isize {
             offset,
         } => {
             syscalls::mm::sys_mmap(addr, length, prot, flags, fd, offset).map_or_else(
-                |e| e.as_neg() as isize,
+                |e| i64::from(e.as_neg()),
                 |ptr| {
-                    let Ok(addr) = isize::try_from(ptr.as_usize()) else {
-                        // Note it assumes user space address does not exceed isize::MAX (0x7FFF_FFFF_FFFF_FFFF).
+                    let Ok(addr) = i64::try_from(ptr.as_usize()) else {
+                        // Note it assumes user space address does not exceed i64::MAX (0x7FFF_FFFF_FFFF_FFFF).
                         // For Linux the max user address is 0x7FFF_FFFF_F000.
                         unreachable!("invalid user pointer");
                     };
@@ -283,10 +283,11 @@ pub fn syscall_entry(request: SyscallRequest<Platform>) -> isize {
             mode,
         } => {
             let Some(path) = pathname.to_cstring() else {
-                return Errno::EFAULT.as_neg() as isize;
+                return i64::from(Errno::EFAULT.as_neg());
             };
-            syscalls::file::sys_openat(dirfd, path, flags, mode).unwrap_or_else(Errno::as_neg)
-                as isize
+            i64::from(
+                syscalls::file::sys_openat(dirfd, path, flags, mode).unwrap_or_else(Errno::as_neg),
+            )
         }
         _ => {
             todo!()
